requirements.txt
streamlit
pymupdf
sentence-transformers
scikit-learn
plotly
pandas
faiss-cpu
textstat